Implemented for OBS Http RESTApi Class.
