+++ UI/obs-http-api.h
@@ -1,0 +1,9 @@
+#pragma once
+
+class OBSHttpApi {
+
+public:
+	bool HttpApiStart();
+	void HttpApiExit();
+	char* GetLocalIpAddr();
+};
